Update devices_autoscan.php
Changed input type for the device_password field from text to password. Currently the password is showing in plain text when you type. This allows a shoulder surfing attack.

Maybe add a option to show the password in plaintext if a user wants to check if they typed it correctly